updated schemas so that geo and time use the same units + resolution types. moved test code from space resolution to tests file
